Resultant-UI
US IT Consulting Services is a Web portal which will provide access US Clients for Software Development Services. There is main 2 parts in this application one for front end second is backend (In- Progress).UI is developed in HTML, CSS, JavaScript where as backend is on Java v1.8 Backend is designed and developed based on Spring Boot Rest application which uses flat files and MySQL data base. User registration followed by login is secured by Spring security based on JWT user authentication and authorization. It is also troublesome tedious to track the invoices for each client on month basis and kept the history of the invoices when the client paid the invoice it will show the transaction the physical asset and track their movement. According the problem discovered above, vulnerabilities arise because of the manual process flow sand its effect on almost all categories of user such as the applicant, an administration office, and staff. The Microsoft Word and Microsoft Excel which currently used to record information of an asset sometimes can be mistaken, incomplete and miscalculate as well as non - integrated.